refactor(applicants): replace ternary chains with useXxxMode hooks + flat switch dispatch
Per jkwon-claude-config #92. Mirrors the DocumentViewer pattern (PR #237/#238)
and the ClassificationRulesPanel pattern (transaction-ternary-dispatch branch).

Refactored components:
- apps/mybookkeeper/frontend/src/app/pages/Applicants.tsx
  → useApplicantsListMode (loading | empty | list) + ApplicantsListBody
- apps/mybookkeeper/frontend/src/app/pages/ApplicantDetail.tsx
  → useApplicantDetailMode (loading | content | null) + ApplicantDetailBody
- apps/mybookkeeper/frontend/src/app/features/applicants/LinkedLeaseDocuments.tsx
  → useLinkedLeaseDocumentsMode (loading | empty | list) + LinkedLeaseDocumentsBody

Discriminated unions added to shared/types/applicant/:
  applicants-list-mode.ts, applicant-detail-mode.ts, linked-lease-documents-mode.ts

New per-state subcomponents (one per file):
  ApplicantDetailBody, ApplicantsListBody, ApplicantsList, ApplicantsEmptyState
  LinkedLeaseDocumentsBody, LinkedLeaseDocumentsList, LinkedLeaseDocumentsLoading,
  LinkedLeaseDocumentsEmpty

Unit tests (55 passing):
  useApplicantsListMode.test.ts (4 tests)
  useApplicantDetailMode.test.ts (6 tests)
  useLinkedLeaseDocumentsMode.test.ts (4 tests)
  All existing applicant tests (41 tests) unchanged and passing

No behavior or styling changes. tsc --noEmit clean.
