+++ apps/mybookkeeper/frontend/src/app/features/applicants/LinkedLeaseDocumentsEmpty.tsx
@@ -1,0 +1,5 @@
+export default function LinkedLeaseDocumentsEmpty() {
+  return (
+    <p className="text-xs text-muted-foreground">No documents on this lease yet.</p>
+  );
+}
