+++ apps/mybookkeeper/frontend/src/app/features/applicants/LinkedLeaseDocumentsLoading.tsx
@@ -1,0 +1,7 @@
+export default function LinkedLeaseDocumentsLoading() {
+  return (
+    <p className="text-xs text-muted-foreground" data-testid="linked-lease-loading">
+      Pulling documents...
+    </p>
+  );
+}
