+++ apps/mybookkeeper/frontend/src/shared/types/applicant/applicant-detail-mode.ts
@@ -1,0 +1,5 @@
+/**
+ * Discriminated union for what ApplicantDetailBody should render.
+ * Replaces a chain of nested ternaries with a single switch.
+ */
+export type ApplicantDetailMode = "loading" | "content";
