+++ apps/mybookkeeper/frontend/src/shared/types/applicant/applicants-list-mode.ts
@@ -1,0 +1,5 @@
+/**
+ * Discriminated union for what ApplicantsListBody should render.
+ * Replaces a chain of nested ternaries with a single switch.
+ */
+export type ApplicantsListMode = "loading" | "empty" | "list";
