+++ apps/mybookkeeper/frontend/src/shared/types/applicant/linked-lease-documents-mode.ts
@@ -1,0 +1,5 @@
+/**
+ * Discriminated union for what LinkedLeaseDocumentsBody should render.
+ * Replaces a chain of nested ternaries with a single switch.
+ */
+export type LinkedLeaseDocumentsMode = "loading" | "empty" | "list";
