+++ apps/mybookkeeper/scripts/test-map.json
@@ -524,7 +524,10 @@
         "SensitiveDataUnlock.test.tsx",
         "Applicants.test.tsx",
         "PromoteFromInquiryPanel.test.tsx",
-        "ContractDatesEditor.test.tsx"
+        "ContractDatesEditor.test.tsx",
+        "useApplicantsListMode.test.ts",
+        "useApplicantDetailMode.test.ts",
+        "useLinkedLeaseDocumentsMode.test.ts"
       ],
       "e2e_specs": [
         "applicants.spec.ts",
